ORG import the heat_diffusion experiment as a (demo) solver and adjust accordingly
The heat_diffusion experiment is meant as a demonstration of how to add a new
solver to the code, and how to refer to it in a new experiment.

The experiment also demonstrates that guard zone loading can keep up, even when
the solver does of order 50 million cell updates per second.

This commit also makes all experiment compile OK on the iac branch, in preparation
for rebasing the master branch to this more up-to-date state.