DOC upated documentation, with brief descriptions of the demo experiments
Notably, the experiments now contain the very simple SOLVER=heat_diffusion,
and a correspondingly simple experiment that uses it.

This should be seen as a template of how to install a new solver (any solver)
in the DISPATCH code framework, and then setup an experiment that uses it.

Such an addition requires, in the simplest case, only two *.f90 files (the
solver definition and the experiment definition), the two corresponding
Makefiles, and the input.nml with default input parameters

experiments/